Create exercise.swift
When exercising it’s really helpful to have a goal in mind and a plan to achieve that goal. While everyone’s goals are different, we can use structures to devise a plan and keep us on track!

We’ll first create a structure for our exercises. We’ll use this structure to keep track of:

its name.
what muscle groups it targets.
how many repetitions (reps) we should do in one go.
how many rounds of repetitions we should do (sets).
How many total reps we’ll be doing.
Another structure we’ll need to keep track of all these exercises is for our regimen! This structure will:

track the day of the week.
store the exercises we should do for a given day.
include a method to remind us of a given day’s workout regimen.